fix uint

+++ cmd/mqttstorm/main.go
@@ -14,7 +14,7 @@ func main() {
 	broker := flag.String("broker", "mqtt://127.0.0.1:1883", "URI(tcp://{ip}:{port}) of MQTT broker (required)")
 	username := flag.String("username", "admin", "Username for connecting to the MQTT broker")
 	password := flag.String("password", "admin", "Password for connecting to the MQTT broker")
-	clientNum := flag.Uint64("c", 1, "Number of clients")
+	clientNum := flag.Int("c", 1, "Number of clients")
 	flag.Parse()
 
 	if *broker == "" {
@@ -29,6 +29,6 @@ func main() {
 		Broker:   *broker,
 		Username: *username,
 		Password: *password,
-	}, *clientNum)
+	}, int32(*clientNum))
 	mss.ListenAndServe()
 }
